api-tests/requirements.txt
# API Testing Dependencies
pytest==8.4.2
requests==2.32.5
python-dotenv==1.1.1
allure-pytest==2.15.0
allure-python-commons==2.15.0
